kernel: optimize mouse cursor redraw with dirty-rect local updates and UHCI USB tablet driver
- Add local dirty-rectangle cursor redrawing in kernel/gfx_desktop.c and kernel/gfx_fb.c (restore/save background, update LFB directly) to eliminate lag from full scene blitting.
- Add interrupt locking (`cli`/`sti`) specifically around short cursor framebuffer updates to prevent IRQ vs GUI thread race conditions without blocking IRQs during scene drawing.
- Add USB UHCI Tablet driver in kernel/input/usb_tablet.c with non-blocking SET_ADDRESS and SET_CONFIGURATION setup transfers for absolute positioning.
- Fix include paths and uintptr_t declarations for standalone freestanding 32-bit compilation.
- Include -Ikernel in Makefile CFLAGS and update test scripts for usb_tablet.c.
- Retain PS/2 relative mouse as fallback.

+++ kernel/input/usb_tablet.c
@@ -2,6 +2,8 @@
 #include "../pci.h"
 #include "../gfx_fb.h"
 #include "../gfx_desktop.h"
+#include <stdint.h>
+#include <stddef.h>
 
 extern unsigned char inb(unsigned short port);
 extern void outb(unsigned short port, unsigned char value);
@@ -73,17 +75,17 @@ static int uhci_control_transfer(uint8_t dev_addr, uint8_t req_type, uint8_t req
     g_setup_pkt.wIndex = idx;
     g_setup_pkt.wLength = 0;
 
-    g_ctrl_td[0].link = (uint32_t)(uintptr_t)&g_ctrl_td[1] | UHCI_TD_LINK_DEPTH;
+    g_ctrl_td[0].link = (uint32_t)(uint32_t)&g_ctrl_td[1] | UHCI_TD_LINK_DEPTH;
     g_ctrl_td[0].status = 0x80000000 | (3 << 27);
     g_ctrl_td[0].token = (7 << 21) | (0x00 << 19) | (0 << 15) | ((uint32_t)dev_addr << 8) | 0x2D;
-    g_ctrl_td[0].buffer = (uint32_t)(uintptr_t)&g_setup_pkt;
+    g_ctrl_td[0].buffer = (uint32_t)(uint32_t)&g_setup_pkt;
 
     g_ctrl_td[1].link = UHCI_TD_LINK_TERM;
     g_ctrl_td[1].status = 0x80000000 | (3 << 27);
     g_ctrl_td[1].token = (0x7FF << 21) | (0x01 << 19) | (0 << 15) | ((uint32_t)dev_addr << 8) | 0x69;
     g_ctrl_td[1].buffer = 0;
 
-    g_qh.element = (uint32_t)(uintptr_t)&g_ctrl_td[0];
+    g_qh.element = (uint32_t)(uint32_t)&g_ctrl_td[0];
 
     for (int timeout = 0; timeout < 50; timeout++) {
         if (!(g_ctrl_td[1].status & 0x80000000)) {
@@ -149,10 +151,10 @@ void usb_tablet_init(void) {
     g_qh.element = UHCI_TD_LINK_TERM;
 
     for (i = 0; i < 1024; i++) {
-        g_frame_list[i] = (uint32_t)(uintptr_t)&g_qh | 0x02;
+        g_frame_list[i] = (uint32_t)(uint32_t)&g_qh | 0x02;
     }
 
-    outl(g_io_base + UHCI_FLBASEADD, (uint32_t)(uintptr_t)g_frame_list);
+    outl(g_io_base + UHCI_FLBASEADD, (uint32_t)(uint32_t)g_frame_list);
     outw(g_io_base + UHCI_FRNUM, 0);
     outw(g_io_base + UHCI_CMD, UHCI_CMD_RUN | UHCI_CMD_MAXP);
 
@@ -170,9 +172,9 @@ void usb_tablet_init(void) {
     g_td.link = UHCI_TD_LINK_TERM;
     g_td.status = 0x80000000 | (3 << 27);
     g_td.token = (7 << 21) | (0x00 << 19) | (1 << 15) | ((uint32_t)g_dev_addr << 8) | 0x69;
-    g_td.buffer = (uint32_t)(uintptr_t)g_report_buf;
+    g_td.buffer = (uint32_t)(uint32_t)g_report_buf;
 
-    g_qh.element = (uint32_t)(uintptr_t)&g_td;
+    g_qh.element = (uint32_t)(uint32_t)&g_td;
 
     g_tablet_present = 1;
     print_string_serial("USB Tablet: Controller UHCI initialise et enumere\n");
@@ -213,7 +215,7 @@ void usb_tablet_poll(void) {
     g_toggle ^= 1;
     g_td.status = 0x80000000 | (3 << 27);
     g_td.token = (7 << 21) | ((uint32_t)g_toggle << 19) | (1 << 15) | ((uint32_t)g_dev_addr << 8) | 0x69;
-    g_qh.element = (uint32_t)(uintptr_t)&g_td;
+    g_qh.element = (uint32_t)(uint32_t)&g_td;
 }
 #else
 void usb_tablet_init(void) {}
